# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Essential HICS API
  description: |
    Hospital Incident Command System — Phase 1 REST API.
    Covers authentication, RBAC, organizational structure, and audit logging.
  version: 1.0.0
  contact:
    name: HICS Support
    email: support@apexhealth.example
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.hics.yourhospital.org
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ValidationError:
      type: object
      properties:
        message: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    AuthUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        displayName: { type: string }
        mustChangePassword: { type: boolean }
        mfaEnabled: { type: boolean }
    LoginResponse:
      type: object
      properties:
        accessToken: { type: string, description: "JWT — store in memory only" }
        user: { $ref: '#/components/schemas/AuthUser' }
        mfaRequired: { type: boolean }
        userId: { type: string, format: uuid, description: "Only present when mfaRequired=true" }
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Login successful or MFA required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
          headers:
            Set-Cookie:
              description: HttpOnly refresh token cookie (hics_refresh)
              schema: { type: string }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ValidationError' } } } }
        '401': { description: Invalid credentials or account locked }
        '429': { description: Rate limit exceeded }
  /auth/mfa/verify:
    post:
      tags: [Authentication]
      summary: Complete login by verifying TOTP or backup code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, code]
              properties:
                userId: { type: string, format: uuid }
                code: { type: string }
                isBackupCode: { type: boolean, default: false }
      responses:
        '200':
          description: MFA verified, full session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401': { description: Invalid MFA code }
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate refresh token, get new access token
      description: Reads the HttpOnly refresh cookie. Rotates the token on every call.
      security: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
        '401': { description: Refresh token missing, invalid, or expired }
...
  /auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200': { description: Always 200 — email sent if address is registered }
  /auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset password with token from email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 12 }
      responses:
        '200': { description: Password reset successfully }
        '400': { description: Invalid/expired token or weak password }
  /auth/change-password:
    post:
      tags: [Authentication]
      summary: Change password (authenticated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 12 }
      responses:
        '200': { description: Password changed; other sessions revoked }
        '401': { description: Current password incorrect }
...